changed shelter charity register field to big integer field as regular int field isn't big enough

+++ crowdfunding/projects/models.py
@@ -13,7 +13,7 @@ class Shelter(models.Model):
     name = models.CharField(max_length=200)
     description = models.TextField()
     address = models.CharField(max_length=200)
-    charityregister = models.IntegerField()
+    charityregister = models.BigIntegerField()
     is_approved = models.BooleanField(default=False)
     species = models.ManyToManyField(
         PetTag,
